feat: Add trending and popular movie slides to Explore page
This commit adds two new movie slides, "Trending" and "Popular", to the Explore page. The slides are implemented using the MovieSlide component and are fetched from the "/populartv" and "/trending" endpoints respectively. This enhances the functionality and design of the Explore page, providing users with more movie options to explore.